Move Trailer Files to separate subFolder
