+++ src/Kinepolis/Trailer/TrailerRepository.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace CultuurNet\UDB3\Kinepolis;
+namespace CultuurNet\UDB3\Kinepolis\Trailer;
 
 use CultuurNet\UDB3\Model\ValueObject\MediaObject\Video;
